scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
aiohttp==3.9.3
asyncio==3.4.3
pandas==2.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
deep-translator==1.11.4   # Google Translate (繁體中文) — no key needed
wikitextparser==0.56.3     # Parse Wiktionary wikitext
tqdm==4.66.2               # Progress bars
tenacity==8.2.3            # Retry on rate limit
xlrd==2.0.1                # Read Lexique .xls
openpyxl==3.1.2            # Read Lexique .xlsx
langdetect==1.0.9
ujson==5.9.0
aiofiles==23.2.1
